rename(route-folder): 채팅리스트 const 컨벤션 적용
