+++ src/app/(route)/chat/_components/CHATLIST_CONST.ts
@@ -23,3 +23,11 @@ export const FilTER_DROPDOWN_OPTIONS = [
     keyName: "type",
   },
 ] as const;
+
+export const SELECTED_TEXT = {
+  oldest: "오래된순",
+  latest: "최신순",
+  all: "발견/분실",
+  found: "발견",
+  lost: "분실",
+} as const;
