--- src/app/(route)/chat/constants/SELECTED_TEXT.ts
@@ -1,7 +1,0 @@
-export const SELECTED_TEXT = {
-  oldest: "오래된순",
-  latest: "최신순",
-  all: "발견/분실",
-  found: "발견",
-  lost: "분실",
-} as const;
